+++ functions.php
@@ -78,6 +78,7 @@ get_template_part( 'lib/meta/meta-boxes-page-newsletters' );
 get_template_part( 'lib/meta/meta-boxes-taxonomy' );
 get_template_part( 'lib/meta/meta-boxes-category-tyskysour' );
 get_template_part( 'lib/meta/meta-boxes-posttype-job' );
+get_template_part( 'lib/meta/meta-boxes-posttype-contributor' );
 
 get_template_part( 'lib/theme-options' );
 get_template_part( 'lib/options-front-page' );