+++ partials/singles/articles/articles-header-basic.php
@@ -5,15 +5,15 @@
 <div class="flex-grid-row margin-bottom-basic">
   <div class="flex-grid-item flex-item-s-12 flex-offset-s-0 flex-item-m-7 flex-offset-m-1 flex-item-l-6 flex-offset-l-1 flex-item-xl-5 flex-offset-xl-1 flex-item-xxl-5 flex-offset-xxl-1 margin-bottom-basic">
     <h1 id="single-articles-title" class="margin-bottom-small"><?php the_title(); ?></h1>
-    
+
     <?php
       if (!empty($meta['_cmb_standfirst'])) {
     ?><h2 class="margin-bottom-small js-fix-widows"><?php echo $meta['_cmb_standfirst'][0]; ?></h2>
     <?php
       }
     ?>
-    
-    <?php get_template_part('partials/singles/articles/components/articles-header-author'); ?>
+
+    <h3>by <?php render_bylines($post->ID, true); ?></h3>
     <h3><?php the_time('j F Y'); ?></h3>
   </div>