+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.7.0] - 2022-07-01
+
+### Added
+
+- Contributors post type and association to posts
+- Auto short bio on articles with associated contributor
+
+### Changed
+
+- Foreign Agent archive copy changes
+
+### Fixed
+
+- Foreign Agent query hook targeting
+
 ## [3.6.1] - 2022-05-24
 
 ### Changed